refactored auth module to follow style guide with each component in separatedirectory
