+++ src/app/app.component.html
@@ -1,10 +1,7 @@
 <!-- <ul>
   <li *ngFor="let user of users">{{ user.firstname }}</li>
 </ul> -->
-
-<app-header></app-header>
-<div class="row">
-	<div class="col-md-8">
-		<router-outlet></router-outlet>
-	</div>
+<div class="container">
+	<app-header></app-header>
+	<router-outlet></router-outlet>
 </div>
