+++ src/app/auth/logout/user-logout.component.html
@@ -1,0 +1,1 @@
+<button class="btn btn-danger" (click)="onLogout">Logout</button>
