+++ src/app/auth/logout/user-logout.component.ts
@@ -12,4 +12,8 @@ export class UserLogoutComponent implements OnInit {
   ngOnInit() {
   }
 
+  onLogout() {
+  	
+  }
+
 }
