+++ src/app/auth/signup/user-signup.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormControl, Validators} from '@angular/forms';
 
-import { User } from './user.model';
-import { AuthService } from './auth.service';
+import { User } from '../user.model';
+import { AuthService } from '../auth.service';
 
 
 function emailDomainValidator(control: FormControl) {
