--- src/app/auth/user-logout.component.html
@@ -1,1 +1,0 @@
-<button class="btn btn-danger">Logout</button>
