+++ src/app/messages/message-list/message-list.component.html
@@ -1,8 +1,12 @@
-<div class="col-md-8">
-	<div class="panel panel-default">
-	  <div class="panel-heading">Message List</div>
-	  <div class="panel-body">
-			<app-message [message]="message" *ngFor="let message of messages"></app-message>
+<div class="container-fluid">
+	<div class="row">
+		<div class="col-md-12">
+			<div class="panel panel-default">
+			  <div class="panel-heading">Message List</div>
+			  <div class="panel-body">
+					<app-message [message]="message" *ngFor="let message of messages"></app-message>
+				</div>
+			</div>
 		</div>
 	</div>
 </div>
