+++ src/app/messages/message/message.component.html
@@ -7,7 +7,7 @@
   </div>
   <footer class="panel-footer">
   	<div class="info">
-  		<span class="author">author</span> <span class="timestamp">{{ message.dateTime }}</span>
+  		<span class="author">Natasha</span> <span class="timestamp">{{ message.dateTime }}</span>
   	</div>
 		<div class="actions">
 			<a href="#">edit</a>
